+++ src/main/java/de/samply/blaze/BlazeStoreClient.java
@@ -37,42 +37,6 @@ public class BlazeStoreClient {
     return fetchPatientBundle(pseudonym);
   }
 
-  private void doSomething(Bundle bundle){
-    Resource patient = getFirstPatient(bundle);
-    String attribute = fetchAttribute("Patient.gender.value", patient);
-    Resource histologie = getFirstHistologie(bundle);
-    String attribute2 = fetchAttribute("Observation.value.coding.code", histologie);
-    String attribute3 = fetchAttribute("Observation.value.coding.code", patient);
-    String attribute4 = fetchAttribute("Bundle.entry.select(resource as Patient).gender.value", bundle);
-    String attribute5 = fetchAttribute("Bundle.entry.select(resource as Observation).where(code.coding.code = '59847-4').value.coding.code", bundle);
-    String attribute6 = fetchAttribute("Bundle.entry.select(resource as Observation).subject.reference.value", bundle);
-  }
-
-  private String fetchAttribute(String fhirPath, Resource resource){
-    ExpressionNode genderNode = fhirPathEngine.parse(fhirPath);
-    List<Base> baseList = fhirPathEngine.evaluate(resource, genderNode);
-    return (baseList != null && baseList.size() > 0) ? baseList.get(0).toString() : null;
-  }
-
-  private Resource getFirstPatient(Bundle bundle){
-    for (BundleEntryComponent component : bundle.getEntry()){
-      if (component.getResource() instanceof Patient){
-          return component.getResource();
-      }
-    }
-    return null;
-  }
-
-  private Resource getFirstHistologie(Bundle bundle){
-    for (BundleEntryComponent component : bundle.getEntry()){
-      if (component.getResource() instanceof Observation &&
-          component.getResource().getMeta().getProfile().get(0).toString().contains("Histologie")){
-        return component.getResource();
-      }
-    }
-    return null;
-  }
-
   private Bundle fetchPatientBundle(String patientPseudonym) {
     return client.search()
         .byUrl("Patient?identifier=" + patientPseudonym)