+++ CHANGELOG.md
@@ -4,5 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2022-11-01]
+## [1.0.0 - 2022-12-09]
 ### Added
+- Search Controller
+- Blaze Store Client
+- LDM Client
+- Containers and Containers Templates
+- FHIR-Bundle to Containers Converter
+- FHIR Rev Includes in Containers Templates